requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.11.4
pydantic-settings==2.2.1

# HTTP Client (async)
httpx[http2]==0.28.1

# OpenAI
openai==1.77.0

# Redis (async)
redis==5.0.1

# JSON (fast serialization)
orjson==3.10.12

# Environment
python-dotenv==1.0.1

# Typing
typing-inspection==0.4.0
typing_extensions==4.13.2
annotated-types==0.7.0

# Utils
PyYAML==6.0.2

# Database (PostgreSQL)
sqlalchemy[asyncio]==2.0.35
psycopg2-binary==2.9.10
asyncpg==0.30.0
alembic==1.14.0

# PDF Processing
pdf2image==1.17.0
Pillow==10.4.0

# JWE Authentication
jwcrypto==1.5.6

# Fernet encryption for stored credentials
cryptography>=44.0.0

# OpenTelemetry
opentelemetry-distro==0.50b0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-httpx==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-instrumentation-redis==0.50b0
opentelemetry-instrumentation-logging==0.50b0
